Using babel-node and reorganized directories
