--- backend/api/post/post.route.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const postController = require('./post.controller')
-
-const router = express.Router()
-
-router.route('/').get(postController.listConfirmed)
-
-module.exports = router
