+++ backend/src/index.js
@@ -1,13 +1,13 @@
 require('dotenv').config()
 const app = require('express')()
 const cors = require('cors')
-const demux = require('./demux')
-const postRoutes = require('./api/post/post.route')
+const demux = require('./services/demux')
+const postsRoutes = require('./api/posts')
 const io = require('./utils/io')
 
 app.use(cors())
 
-app.use('/posts', postRoutes)
+app.use('/posts', postsRoutes)
 
 const server = app.listen(process.env.PORT, () => console.info(`Example app listening on port ${process.env.PORT}!`))
