+++ backend/src/models/index.js
@@ -1,0 +1,7 @@
+import BlockIndexState from './block-index-state.model';
+import Post from './post.model';
+
+export {
+    BlockIndexState,
+    Post
+}
