+++ backend/src/services/demux/updaters/blog/editpost.js
@@ -1,4 +1,5 @@
 async function editPost (state, payload, blockInfo, context) {
+  console.log(payload)
   try {
     await state.post.findByIdAndUpdate({ timestamp: payload.data.timestamp, author: payload.data.author }, payload.data).exec()
   } catch (err) {
